update test and ci
